HW Assignment: Higher-Order Functions and Objects
This assignment is graded based on correctness and will require you to use higher-order functions to sort automobiles. The description is below and can also be found here  (Links to an external site.)Links to an external site.on jsFiddle. You should submit a single .js file called automobile.js which when run with node.js using the command "node automobile.js" produces the described results. You must make use of  higher-order functions to sort the cars. You should not, for example, create entirely separate functions each with dedicated loops to sort the cars. You will need a loop (or potentially more than one loop depending on your sorting algorithm of choice) in the sortArr function but that is pretty much it.