HW Assignment: DOM and Events
For this assignment you should submit a single zip file that contains the following two files:

index.html
script.js
index.html should be a skeleton HTML page. So it should have the following tags:

doctype
html
head
meta
title
body
script
If you were to open index.html without including the associated JavaScript it should be entirely blank.

You should then use JavaScript to create all of the content of this page and append it to the body of the page. That content should include:

A 4x4 table
The top row should be a header row with header cells
The 4 header cells, from left to right should say "Header 1", "Header 2" ... "Header 4
The non header cells should contain their position. The upper left cell should contain the text "1, 1", the cell to its right "2, 1", the cell below it "1, 2" and so on.
4 directional buttons (up, down, left right)
A button labeled "Mark Cell"
When the page is loaded the upper left, non-header cell of the table should be 'selecte